+++ lib/ReactTestUtils.js
@@ -17,8 +17,8 @@
     //收集符合条件的元素节点
     function findChildren(fiber, callback, ret) {
 
-        for (var i in fiber._children) {
-            var a = fiber._children[i];
+        for (var i in fiber.children) {
+            var a = fiber.children[i];
             findAllInRenderedTreeImpl(a, callback, ret);
         }
     }
