+++ lib/createClass.js
@@ -1,5 +1,5 @@
 /**
- * React.createClass补丁 by 司徒正美 Copyright 2018-05-05
+ * React.createClass补丁 by 司徒正美 Copyright 2018-05-08
  * IE9+
  */
