+++ lib/createReactNoop.js
@@ -6,14 +6,14 @@ function collectVDOM(vdom, array) {
             children: children
         };
         array.push(element);
-        var c = vdom._children;
+        var c = vdom.children;
         for (var i in c) {
             collectVDOM(c[i], children);
         }
     } else if (vdom.tag === 6) {
         array.push(vdom.text + "");
     } else {
-        var c = vdom._children;
+        var c = vdom.children;
         for (var i in c) {
             collectVDOM(c[i], array);
         }
