+++ lib/shallowCompare.js
@@ -1,5 +1,5 @@
 /**
- * shallowCompares补丁 by 司徒正美 Copyright 2018-05-05
+ * shallowCompares补丁 by 司徒正美 Copyright 2018-05-08
  * IE9+
  */
